organize spec helpers
