--- lib/jets/spec_helpers/params.rb
@@ -1,10 +1,0 @@
-module Jets
-  module SpecHelpers
-    class Params
-      attr_accessor :path_params, :body_params, :query_params
-      def initialize(path_params={}, body_params={}, query_params={})
-        @path_params, @body_params, @query_params = path_params, body_params, query_params
-      end
-    end
-  end
-end
